WARNING: Read the commit message. Infra dir moved.
As we do not have a way to force the update/sync of the cookbook with chef-provisioning we want to move the `infra/` directory out from `.chef/local-mode-cache/cache/infra/` so we can sync the cookbook with just clearing the cache `rm -rf .chef/local-mode-cache` and running `chef-client -z` again without any harm of loosing Delivery/Chef Server Data.

We are put it besides the trusted_certs directory. (.chef/delivery-cluster-data)

WARNING: This will break existing implementations or running environments so you have to move the infra folder and re name it to `delivery-cluster-data` before using this versioin of delivery-cluster.

Run:
```
$ cd delivery-cluster
$ cp -r .chef/local-mode-cache/cache/infra .chef/delivery-cluster-data
```