Allow the Artifact pass-through the delivery server
Added `pass-through` parameter that will allow the Artifact pass-through the delivery server.

By default we will set this to `true` but if your delivery server does NOT have access to Chef VPN
you have to set this to `false` so that the artifact can be downloaded locally and then upload it to
the delivery server.

NOTE: This feature requires that the delivery server has VPN Access.

This commit also substitute the `md5` checksum for `sha256` that is the one
we need for the `remote_file` resource within the `delivery-server` cookbook.