Abstract the specific configurations by providers (aws)
There will be a time where we are going to be able to provision `delivery-cluster`
in different providers. This is the starting poing to do so. Abstracting the settings
will let us manage them separate from the actual machine resources.

This is also useful when other cookbooks depend on `delivery-cluster`
and they need to configure the same set of settings. e.g. (delivery-demo)