Splunk Integration
Now you are going to be able to provision your own Splunk Server
and it will be automatically waiting for any Analytics information
since we deploy the Analytics Splunk App on it. Yeah!

Analytics Server now detects when there is a Splunk Server if it
so, it will automatically configure `features['integration'] = true`
into the `/etc/opscode-analytics/opscode-analytics.rb`.

The way you will provision this is by running:

```
$ bundle exec chef-client -z -o delivery-cluster::setup_splunk -E chefconf
```

Additionally dependencies like `chef-vault` & `chef-splunk`. But missing documentation.