Provisioning Driver Abstraction
This commit is abstracting the Provisioning Drivers, at the moment
this will support `AWS` and `SSH` but in the future if we will want
to use `Vagrant` we just create anothe abstration from the `Base`
class.

Deleted the file `_aws_settings.rb` that holded specific attributes
for `AWS` now it is coming from the Abstraction.

Modified a buch of recipes to be in compliant to the new `_settings.rb`
recipe file.