Add support for Delivery license key
An upcoming release of Delivery will require installations to include a valid
license key. This change introduces a number of changes to make it so that this
license key can be processed and uploaded to the Delivery Server as part of the
normal delivery-cluster run. These changes include:
  * During the compile phase, make sure that the license key is present on the
  provisioning node. This will prevent a user from getting deep into the
  setup process only to have it fail.
  * If we cannot find the license key, provide helpful information to the user
  as to how to remediate the issue.
  * Perform the necessary key processing on the provisioner and send only the
  necessary files (delivery-license and delivery-license.sig) up to the server.

I documented in the README the appropriate process for procuring the license key
and specifying that key's path on the provisioner using your environment file.

In addition to these changes I also:
  * incorporated the Makefile from `delivery-demo` into `delivery-cluster`
  (making the appropriate changes to the README).
  * Updated Berksfile.lock and Gemfile.lock