Do not error on missing local provisioner
There may be situations where users will use Chef to pull down the
license file. This means it won't be available in the compile phase. As
such, we will trust that if the user has set the license path they know
what they are doing and will _not_ error if we can't find the license
file on the provisioner.