Rename methods to say FQDN instead of IP
Lets honor the methods names with the actual thing they are doing.

It is very confusing that the methods says `chef_server_ip` when in
reality it is getting the `fqdn` and if that doesn't exist then we
get the ip address.

This change also prepares another change to fix:
https://github.com/opscode-cookbooks/delivery-cluster/issues/94