Change chef-zero port to 8890 & enable opscode-reporting
Since some of our `chef-server` plugins use `chef-zero` to reconfigure
there is a port conflict. This PR changes the default port from `8889`
to `8890` to fix this problem. We would be able to remove this in the
future once it no longer conflicts.

We are enabling `opscode-reporting` by default.

Fixing issue https://github.com/opscode-cookbooks/delivery-cluster/issues/79

**If users are using `chef-provisioning-ssh` they might need to remove
`.chef/provisioning` directory in order to apply this change**