Common Cluster Recipes & pkg_repo_management recipe
This commit is including a new attribute called `common_cluster_recipes`
that will let you add any recipe that you need to add to the run_list of
all the servers of the delivery-cluster.

As an example:
* We would like to aply a security policy to every single server on the cluster.

  `security_policies::lock_root_login` locks down root login

This attribute would looks like:

```
default['delivery-cluster']['common_cluster_recipes'] = ['security_policies::lock_root_login']
```

Additionally it added a recipe `pkg_repo_management` that will do the
management of the package repositories in a cross-platform manner.

As reference we have a new helper method that is called `default_cluster_recipes`
that will Return the default cluster recipes from Chef. These recipes are the ones
we internally need to let delivery-cluster work properly.