AWS pipeline stabilization
As a developer that commits code to the delivery-cluster project
I expect the pipeline to be stable and consistent on every patchset
so I can push my code without re-running the phases/stages multiple times.

**Problem Description**

Currently the delivery-cluster pipeline is very unstable, as we introduced
the matrix-build (red-pill) we are running it multiple times making the
errors to show aleatory on different pipelines. The errors are coming from
different places like push-jobs, chef-provisioning, and other components of
delivery-cluster. The Workaround in place is to re-run the automation
(rake setup:cluster) multiple times before marking it as failed.
But what we really want is to have a clean unique run.

**Acceptance Criteria**

* Document the functionality of the delivery-cluster pipeline: This will give
more visibility to other developers that comes to the project for the first time.
* Translate to `bash` scripts into Ruby code: This will make the pipeline more robust.
* Gather information of chef-provisioning issues and report them: Instead of
re-running multiple times, we would like to get every single problem with chef-provisioning
and report it so the team can fix it.