Ensure all files get uploaded to the build-nodes
The `machine` resource in `chef-provisioning` does not allow to have
`files` and `file` statements together. That means that one of them will
take precedence.

This commit fixes the problem by abstracting the files to upload into a
library in `libraries/helpers_builders.rb` that will lazy load to ensure
that we upload all the files (even the ones that get render on the
chef-zero run) to the build-nodes

Closes https://github.com/chef-cookbooks/delivery-cluster/issues/164