Include Supermarket in setup cluster
If we are Implementing Delivery for Cookbooks Workflow we have to have a
Supermarket Server. Before we had to run the `rake setup:cluster` and
then `rake setup:supermarket`.

This commit adds Supermarket to the default rake task for setting up the
Delivery Cluster (`rake setup:cluster`) so you dont have to run two
commands.

Additionally it verifies if Supermarket Server is enabled, if it is not
then, it will skip the recipe.