Wrap activate_supermarket in a ruby_block
Now that Supermarket is part of the default setup of Delivery. When ever
we compile the recipe `setup_supermarket` it activates Supermarket
inmediately. This is comflicting with the logic we have on the recipe so
we have to wrap the `activate_supermarket` method into a `ruby_block`
resource so we delay the activation of Supermarket. :smile: