Fix Delivery pipeline
This is a fix to our Delivery pipeline, when we introduced multiple
platforms to our environment we now have to be carefully with the
version we run on. Specially to run on the same node in Provision and
Deploy. Otherwise we will end up with an unstable pipeline.

An example to this is the difference of versions that conflicts with
other pipeline runs:
```
  * delivery_rspec_block[Smoke Test Delivery Cluster] action run

  1)  Delivery Server deployed version is 0.3.390
     Failure/Error: expect(@browser.text).to include("delivery #{delivery_details['version']}")
       Diff:
       @@ -1,2 +1,60 @@
       -delivery 0.3.390
       +delivery 0.3.374
```

This commit match the Provision and Deploy phases to an speficic node.