Remove packagecloud dependency in favor of chef-ingredient
We where using a custom `packagecloud` cookbook for the installation of
delivery, this commit removes that dependency and introduce
`chef-ingredient` instead which is our new standard to install Chef
Software.

As this is changing the way we install/upgrade delivery we will bump the
major version of this cookbook.

Addidionally updated the cookbooks versions to the latest.