Get rid of `bundle` in favor of ChefDK
We have seen many problems by maintaining the `Gemfile.lock` therefore
this commit is removing `bundle` from this cookbook and now we are going
to relay in `ChefDK` 100%. The process for many customers/end-users will
change a little bit so we are bumping the major to `0.5.0`.

The recomended chefdk version is always the latest and any new fix will
be consume from there.