Vefiry ChefDK version and Install Prerequisites
We will be enforcing the use of ChefDK `0.10.0` and also we are
introducing the recipe `provisioner.rb` which will have all the
prerequisites we need to have on the provisioning node. For now the only
things that are missing are the provisioning drivers.

Keep in mind sometimes to install gems we need to run it as root,
probable the end-user might need to run the prerequisites like:
```
$ sudo rake setup:prerequisites
```

Closes https://github.com/chef-cookbooks/delivery-cluster/issues/183