Customization of Chef Server user password
This commit enables the end-user to customize the password of the
delivery user in the Chef Server. If you want to modify the password you
just need to change the attribute:
```
node['delivery-cluster']['chef-server']['delivery_password'] = 'MySecurePassword'
```

If the password is not customized `delivery-cluster` will automatically
generate a random one.

Closes https://github.com/chef-cookbooks/delivery-cluster/issues/192