Build-Cookbook fix to save chef-server url
The build-cookbook gathers the information of the cluster through the
Rake task `rake info:list_core_services` and the last change modified
the behavior so now here we are fixing how we read the data.

Additionally added a quick refactory of the code to clean it a little
bit. Tested locally the functionality:

```
[53] pry(main)> node.run_state['delivery']
=> {"stage"=>
  {"data"=>
    {"cluster_details"=>
      {"delivery"=>{"url"=>"10.194.10.240"},
       "build_nodes"=>["10.10.10.10"],
       "supermarket_server"=>{"url"=>"10.194.12.170"},
       "chef_server"=>{"url"=>"10.194.15.21"}}}}}
```