Enable Delivery to work with OEL
There are two PR's that enables Delivery to work with OEL systems,
one of them is against packagecloud cookbook and it has already been
merged.
https://github.com/computology/packagecloud-cookbook/pull/32

The other one though is still pending. It is to runit but it is working
fine.
https://github.com/hw-cookbooks/runit/pull/164

Until we get this merged we will point to this repo that has the
improvement.
https://github.com/afiune/runit/tree/afiune/make-it-work-in-oel