Remove and ignore Berksfile.lock
We have been dealing with some painful things with customers these week
and all because of the use of the `Berksfile.lock` file. At this point
it is doing more harm than helping us so I think it is time to remove it
and pin the cookbook versions we need to on the actual `metadata.rb`.