Add Insights library
Insights needs to be a component with its own library so that it can
coexist with the other compoponets like the ChefServer since we pass
extra attributes to configure the rabbitmq.

This change is abstracting the functionality and also adding extra
helper functions to know whether or not Insight is enabled.