For existing chef-server disable Supermarket/Analytics
With this commit you can't bring your own chef-server and try to install
Supermarket or Analytics server with delivery-cluster. Why? Because the
in order to install the products we need to modify the config in the chef
server. Il faut faire l'installation manuel.

Additionally add a ruby_block to print the Insights config that the user
needs to add manually to the existing chef-server.

Bump version to 0.6.10