Improve terminal escape parsing in delivery-matrix
delivery-matrix shells out to `delivery review` which uses terminal
escape codes to provide fancy output to the user. These escape codes can
vary based on the value of the TERM environment variable. The previous
code attempted to strip all of the escape codes. The new code attempts
to extract the url assuming that \e (ESC) doesn't appear in a valid URL.